requirements.txt
openai~=1.82.0
requests
Pillow>=9.5.0
SpeechRecognition
pyaudio
# If you encounter issues with installing `pyaudio`, try installing manually using the .whl in `install` folder
# E.g. `pip install .\install\PyAudio-0.2.11-cp37-cp37m-win_amd64.whl`